% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bar_style}
\alias{bar_style}
\title{Create bars for percentages}
\usage{
bar_style(
  width = 1,
  fill = "#D9D9D9",
  height = "1.5em",
  align = c("left", "right"),
  backgroundPosition = "0.25em 0.375em",
  color = "black"
)
}
\arguments{
\item{width}{scalar; width of bar as proportion of cell}

\item{fill}{string; hex color to use for bar}

\item{height}{string; height of bar}

\item{align}{string; one of "left" or "right" to indicate alignment of bar}

\item{backgroundPosition}{string; adjust position of bar}

\item{color}{string; color to use for text}
}
\description{
Create bars for percentages
}
