% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{table_values}
\alias{table_values}
\title{Frequency Table using reactable}
\usage{
table_values(cur_col = "selsibtype", file = getOption("wlsDataPath"))
}
\arguments{
\item{cur_col}{string; name of column to create frequency table from}

\item{file}{string; path to file}
}
\description{
Frequency Table using reactable
}
