% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_variables_table.R
\name{wls_variables_table}
\alias{wls_variables_table}
\title{Table of WLS Variables}
\usage{
wls_variables_table(wls_dat, rstudioapi_available = F)
}
\arguments{
\item{wls_dat}{\code{data.frame} with columns \code{var_name}, \code{visit}, and \code{labels}}

\item{rstudioapi_available}{\code{logical} to indicate if the R package \code{rstudioapi} is installed and available.}
}
\description{
Table of WLS Variables
}
\keyword{internal}
