% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsDataBrowser-package.R
\docType{package}
\name{wlsDataBrowser-package}
\alias{wlsDataBrowser}
\alias{wlsDataBrowser-package}
\title{wlsDataBrowser: WLS Data Browser}
\description{
The Wisconsin Longitudinal Study (https://wls.wisc.edu) is an ongoing longitudinal study first started in 1957. As such, a large number of variables have been collected. This R-package contains a small Shiny application which allows you to browse the variables available in the WLS data. Note: data not included, but freely available from the WLS website.
}
\author{
\strong{Maintainer}: Ralph Møller Trane \email{rmtrane@gmail.com}

}
\keyword{internal}
