% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wlsDataBrowserUI.R
\docType{data}
\name{wlsDataBrowserUI}
\alias{wlsDataBrowserUI}
\title{UI Part of Shiny App}
\format{
An object of class \code{bslib_page} (inherits from \code{shiny.tag.list}, \code{list}) of length 5.
}
\usage{
wlsDataBrowserUI
}
\description{
UI Part of Shiny App
}
\keyword{datasets}
